split qlinear binary family
