+++ tests/test_flatbuffer_direct_architecture.py
@@ -24,6 +24,11 @@ DEPENDENCY_SCOPED_FILES = [
     / "onnx2tf"
     / "tflite_builder"
     / "op_builders"
+    / "qlinear_binary.py",
+    REPO_ROOT
+    / "onnx2tf"
+    / "tflite_builder"
+    / "op_builders"
     / "qlinear_pool.py",
     REPO_ROOT / "onnx2tf" / "tflite_builder" / "pytorch_codegen_utils.py",
     REPO_ROOT / "onnx2tf" / "tflite_builder" / "pytorch_layout_utils.py",
@@ -146,6 +151,39 @@ def test_qlinear_fc_builders_stay_in_family_module() -> None:
     assert {"build_qlinear_matmul_op", "build_qgemm_op"} <= family_functions
     assert "build_qlinear_matmul_op" not in legacy_functions
     assert "build_qgemm_op" not in legacy_functions
+
+
+def test_qlinear_binary_builders_stay_in_family_module() -> None:
+    family_path = (
+        REPO_ROOT
+        / "onnx2tf"
+        / "tflite_builder"
+        / "op_builders"
+        / "qlinear_binary.py"
+    )
+    legacy_path = (
+        REPO_ROOT
+        / "onnx2tf"
+        / "tflite_builder"
+        / "op_builders"
+        / "quantized.py"
+    )
+    family_source = family_path.read_text(encoding="utf-8")
+    legacy_source = legacy_path.read_text(encoding="utf-8")
+    family_functions = {
+        node.name
+        for node in ast.parse(family_source).body
+        if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef))
+    }
+    legacy_functions = {
+        node.name
+        for node in ast.parse(legacy_source).body
+        if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef))
+    }
+
+    expected = {"build_qlinear_add_op", "build_qlinear_mul_op"}
+    assert expected <= family_functions
+    assert expected.isdisjoint(legacy_functions)
 
 
 def test_qlinear_pool_builders_stay_in_family_module() -> None:
