+++ tests/test_flatbuffer_direct_qlinear_fc.py
@@ -1,15 +1,11 @@
 from __future__ import annotations
 
-import hashlib
-import json
-from typing import Any
-
 import numpy as np
 import onnx
 import pytest
 from onnx import TensorProto, helper, numpy_helper
 
-from onnx2tf.tflite_builder.lower_from_onnx2tf import lower_onnx_to_ir
+from tests.flatbuffer_direct_fingerprint import model_ir_fingerprint
 
 
 def _make_qlinear_fc_model(op_type: str) -> onnx.ModelProto:
@@ -61,68 +57,6 @@ def _make_qlinear_fc_model(op_type: str) -> onnx.ModelProto:
     )
 
 
-def _normalize_for_fingerprint(value: Any) -> Any:
-    if isinstance(value, np.ndarray):
-        return {
-            "dtype": str(value.dtype),
-            "shape": list(value.shape),
-            "sha256": hashlib.sha256(value.tobytes()).hexdigest(),
-        }
-    if isinstance(value, np.generic):
-        return value.item()
-    if isinstance(value, dict):
-        return {
-            str(key): _normalize_for_fingerprint(value[key])
-            for key in sorted(value)
-        }
-    if isinstance(value, (list, tuple)):
-        return [_normalize_for_fingerprint(item) for item in value]
-    return value
-
-
-def _model_ir_fingerprint(model: onnx.ModelProto, op_type: str) -> str:
-    model_ir = lower_onnx_to_ir(
-        model,
-        output_file_name=f"{op_type}_family_fingerprint",
-        allow_custom_ops=False,
-    )
-    payload = {
-        "inputs": model_ir.inputs,
-        "outputs": model_ir.outputs,
-        "operators": [
-            {
-                "type": operator.op_type,
-                "inputs": operator.inputs,
-                "outputs": operator.outputs,
-                "options": _normalize_for_fingerprint(operator.options),
-                "version": operator.version,
-                "custom": getattr(operator, "custom_code", None),
-            }
-            for operator in model_ir.operators
-        ],
-        "tensors": {
-            name: {
-                "dtype": tensor.dtype,
-                "shape": tensor.shape,
-                "signature": tensor.shape_signature,
-                "data": _normalize_for_fingerprint(tensor.data),
-                "quant": _normalize_for_fingerprint(
-                    None
-                    if tensor.quantization is None
-                    else vars(tensor.quantization)
-                ),
-            }
-            for name, tensor in sorted(model_ir.tensors.items())
-        },
-    }
-    serialized = json.dumps(
-        payload,
-        sort_keys=True,
-        separators=(",", ":"),
-    ).encode()
-    return hashlib.sha256(serialized).hexdigest()
-
-
 @pytest.mark.parametrize(
     ("op_type", "expected_fingerprint"),
     [
@@ -136,11 +70,11 @@ def _model_ir_fingerprint(model: onnx.ModelProto, op_type: str) -> str:
         ),
     ],
 )
-def test_qlinear_fc_family_preserves_pre_extraction_model_ir_fingerprint(
+def test_qlinear_fc_family_preserves_pre_extractionmodel_ir_fingerprint(
     op_type: str,
     expected_fingerprint: str,
 ) -> None:
-    assert _model_ir_fingerprint(
+    assert model_ir_fingerprint(
         _make_qlinear_fc_model(op_type),
         op_type,
     ) == expected_fingerprint
